LayoutTests/http/tests/in-app-browser-privacy/context-string-preconnect-expected.txt

CONSOLE MESSAGE: Successfully preconnected to http://localhost:8000/
Tests that preconnect tasks get marked with the proper context string.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Context string for preconnect correctly captured.
PASS successfullyParsed is true

TEST COMPLETE
